write variables on exit, cleanup of trailing whitespace
In the process I improved some of the file handling code; zGetUserDir now creates the user data
directory if it doesn't exist already. zFileExists replaced by zPathExists that returns 0 if it
doesn't exist, or otherwise returns wether or not it is a file or directory. zGetPath can now be
forced to only return path for user data dir (i.e. for writing files) using Z_FILE_FORCEUSER flag.