miscellaneous xrandr fixes
Now also grabs X pointer when going fullscreen so I can confine the pointer to the window. This is
to prevent the cursor from getting into 'dead areas' after switching to a lower resolution mode, and
is probably useful too in multimonitor setups. Unfortunately this makes it impossible for the user
to use key combinations like ALT+TAB to interact with the window manager, so I might need to employ
some kind of hack to ungrab the pointer when some special key (i.e. LALT) is pressed, and grab it
again when the window gains focus, but I'm not sure yet if there is a better way to do this.